+++ SDLTest/src/Level_2_Scene.cpp
@@ -50,10 +50,10 @@ void Level_2_Scene::onLoadObjects()
 
 	GemManager::initialize(10, 8, 60.0f, Vector2D(21.0f, -54.0f));
 
-	std::vector<Uint8> r0Cols = { 4, 5 };
-	std::vector<Uint8> r0r6Cols = { 0, 1, 8, 9 };
-	std::vector<Uint8> r1r5Cols = { 0, 9 };
-	std::vector<Uint8> r7Cols = { 0, 1, 2, 7, 8, 9 };
+	std::vector<Uint64> r0Cols = { 4, 5 };
+	std::vector<Uint64> r0r6Cols = { 0, 1, 8, 9 };
+	std::vector<Uint64> r1r5Cols = { 0, 9 };
+	std::vector<Uint64> r7Cols = { 0, 1, 2, 7, 8, 9 };
 
 	GemManager::getInstance()->setBlockedCells(0, r0Cols, true);
 	GemManager::getInstance()->setBlockedCells(0, r0r6Cols, true);
